+++ config/model.go
@@ -26,7 +26,7 @@ type LogConfig struct {
 type BootstrapConfig struct {
 	Prune            bool
 	GenesisAccounts  []models.GenesisAccount `json:"-"`
-	BlocksToFinalise uint64
+	BlocksToFinalise uint32
 	BootstrapNodeURL *string
 	ChainSpecPath    *string
 }
